Add saveBlob in utils/core.js
